@@ -35,86 +35,86 @@
     "dep-graph": "nx dep-graph",
     "help": "nx help",
     "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
-    "snyk-protect": "snyk protect",
-    "prepare": "npm run snyk-protect",
-    "storybook": "yarn docs:json && nx run shared-storybook:storybook",
-    "docs:json": "compodoc -p ./packages/angular/tsconfig.lib.prod.json -e json -d ./packages/storybook",
+    "angular:storybook": "yarn docs:json && nx run angular-storybook:storybook",
+    "docs:json": "compodoc -p ./packages/angular/tsconfig.lib.prod.json -e json -d ./packages/angular-storybook",
     "run-many": "yarn docs:json && npx nx run-many --target=build --configuration=production --all"
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^11.0.7",
-    "@angular/common": "^11.0.7",
-    "@angular/compiler": "^11.0.7",
-    "@angular/core": "^11.0.7",
-    "@angular/forms": "^11.0.7",
-    "@angular/platform-browser": "^11.0.7",
-    "@angular/platform-browser-dynamic": "^11.0.7",
-    "@angular/router": "^11.0.7",
+    "@angular/animations": "^11.2.7",
+    "@angular/common": "^11.2.7",
+    "@angular/compiler": "^11.2.7",
+    "@angular/core": "^11.2.7",
+    "@angular/forms": "^11.2.7",
+    "@angular/platform-browser": "^11.2.7",
+    "@angular/platform-browser-dynamic": "^11.2.7",
+    "@angular/router": "^11.2.7",
     "rxjs": "^6.6.3",
     "tslib": "^2.1.0",
-    "zone.js": "^0.10.3",
+    "zone.js": "^0.11.3",
     "dagre": "^0.8.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.6",
-    "@angular/compiler-cli": "^11.0.7",
-    "@angular/language-service": "^11.0.7",
-    "@angular-devkit/core": "^11.0.6",
-    "@commitlint/cli": "^11.0.0",
-    "@commitlint/config-conventional": "^11.0.0",
+    "@angular-devkit/build-angular": "~0.1102.6",
+    "@angular/compiler-cli": "^11.2.7",
+    "@angular-eslint/eslint-plugin": "~1.0.0",
+    "@angular-eslint/eslint-plugin-template": "~1.0.0",
+    "@angular-eslint/template-parser": "~1.0.0",
+    "@angular/language-service": "^11.2.7",
+    "@angular-devkit/core": "^11.2.6",
+    "@commitlint/cli": "^12.0.1",
+    "@commitlint/config-conventional": "^12.0.1",
     "@babel/core": "^7.12.10",
-    "@nrwl/angular": "^11.1.1",
-    "@nrwl/cli": "11.1.1",
-    "@nrwl/eslint-plugin-nx": "11.1.1",
-    "@nrwl/jest": "11.1.1",
-    "@nrwl/node": "^11.1.1",
-    "@nrwl/tao": "11.1.1",
-    "@nrwl/workspace": "11.1.1",
-    "@nrwl/storybook": "^11.1.1",
-    "@nrwl/cypress": "^11.1.1",
+    "@nrwl/angular": "^11.6.0",
+    "@nrwl/cli": "11.6.0",
+    "@nrwl/eslint-plugin-nx": "11.6.0",
+    "@typescript-eslint/eslint-plugin": "4.3.0",
+    "@typescript-eslint/parser": "4.3.0",
+    "@nrwl/jest": "11.6.0",
+    "@nrwl/node": "^11.6.0",
+    "@nrwl/tao": "11.6.0",
+    "@nrwl/workspace": "11.6.0",
+    "@nrwl/storybook": "^11.6.0",
+    "@nrwl/cypress": "^11.6.0",
+    "@types/resize-observer-browser": "^0.1.5",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/commit-analyzer": "^8.0.1",
     "@semantic-release/exec": "^5.0.0",
     "@semantic-release/git": "^9.0.0",
-    "@semantic-release/npm": "^7.0.8",
-    "semantic-release": "^17.3.1",
-    "cypress": "^6.2.1",
+    "@semantic-release/npm": "^7.0.10",
+    "semantic-release": "^17.4.1",
+    "cypress": "^6.3.0",
     "eslint-plugin-cypress": "^2.11.2",
-    "@types/jest": "26.0.8",
-    "@types/node": "12.12.38",
+    "@types/jest": "26.0.20",
+    "@types/node": "14.14.22",
     "@types/dagre": "^0.7.44",
-    "@typescript-eslint/eslint-plugin": "4.3.0",
-    "@typescript-eslint/parser": "4.3.0",
-    "@storybook/addon-docs": "^6.1.14",
-    "@storybook/addon-essentials": "^6.1.14",
-    "@storybook/addon-links": "^6.1.14",
-    "@storybook/angular": "^6.1.14",
+    "@storybook/addon-docs": "^6.2.1",
+    "@storybook/addon-essentials": "^6.2.1",
+    "@storybook/addon-links": "^6.2.1",
+    "@storybook/angular": "^6.2.1",
     "@compodoc/compodoc": "^1.1.11",
-    "babel-loader": "^8.0.0",
+    "babel-loader": "^8.2.2",
     "react": "^16.0.1",
     "react-dom": "^16.0.1",
     "codelyzer": "^6.0.0",
     "cz-customizable": "^6.3.0",
-    "dotenv": "6.2.0",
-    "eslint": "7.10.0",
-    "eslint-config-prettier": "6.0.0",
+    "dotenv": "8.2.0",
+    "eslint": "7.18.0",
+    "eslint-config-prettier": "7.2.0",
     "git-cz": "^4.7.6",
-    "husky": "^4.3.6",
-    "jest": "26.2.2",
-    "jest-preset-angular": "8.3.1",
-    "ng-packagr": "^11.0.0",
-    "prettier": "2.1.2",
-    "snyk": "1.428.2",
-    "ts-jest": "26.4.0",
+    "husky": "^4.3.8",
+    "jest": "26.6.3",
+    "jest-preset-angular": "8.3.2",
+    "ng-packagr": "^11.1.2",
+    "prettier": "2.2.1",
+    "ts-jest": "26.5.0",
     "ts-node": "~9.1.1",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.3"
+    "tslint": "~6.1.3",
+    "typescript": "~4.1.3"
   },
   "config": {
     "commitizen": {
       "path": "node_modules/cz-customizable"
     }
-  },
-  "snyk": true
+  }
 }
