@@ -3,8 +3,8 @@
   "version": "0.0.1",
   "license": "MIT",
   "peerDependencies": {
-    "@angular/common": "^11.0.5",
-    "@angular/core": "^11.0.5"
+    "@angular/common": "^11.2.7",
+    "@angular/core": "^11.2.7"
   },
   "dependencies": {
     "tslib": "^2.0.0",
