@@ -1,10 +1,13 @@
 /*
  * Public API Surface of @rxzu/angular library
  */
-export * from './lib/defaults';
-export * from './lib/engine.service';
 export * from './lib/diagram/diagram.component';
+export * from './lib/injection.tokens';
+export * from './lib/engine.service';
+export * from './lib/factory.service';
+export * from './lib/registry.service';
 export * from './lib/main.module';
+export * from './lib/defaults/index';
 export * from './lib/utils';
 
 // export all core
