@@ -1,18 +1,22 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { DefaultPortModel } from '@rxzu/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Inject,
+  OnInit,
+} from '@angular/core';
+import { PortModel } from '@rxzu/core';
+import { MODEL } from '../../../injection.tokens';
 
 @Component({
-  selector: 'ngdx-default-port',
+  selector: 'rxzu-default-port',
   templateUrl: './default-port.component.html',
   styleUrls: ['./default-port.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class DefaultPortComponent extends DefaultPortModel implements OnInit {
-  constructor() {
-    super({ type: 'ngdx-default-port' });
-  }
+export class DefaultPortComponent implements OnInit {
+  constructor(@Inject(MODEL) public model: PortModel) {}
 
   ngOnInit() {
-    this.setPainted(true);
+    this.model.setPainted(true);
   }
 }
