@@ -1,9 +1,4 @@
-export * from './default-link/default-link.module';
-export * from './default-node/default-node.module';
-export * from './default-port/default-port.module';
-export * from './default-label/default-label.module';
-
+export * from './default-label/default-label.component';
 export * from './default-link/default-link.component';
 export * from './default-node/default-node.component';
 export * from './default-port/default-port.component';
-export * from './default-label/default-label.component';
