@@ -1,134 +1,203 @@
 import {
-  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
   ElementRef,
+  Inject,
+  HostListener,
   Input,
   NgZone,
   OnDestroy,
   Renderer2,
   ViewChild,
   ViewContainerRef,
+  OnInit,
 } from '@angular/core';
-import { combineLatest, Observable, ReplaySubject } from 'rxjs';
+import { combineLatest, noop, Observable, of } from 'rxjs';
 import { map, switchMap, takeUntil, tap } from 'rxjs/operators';
 import {
-  DiagramEngineCore,
-  DiagramModel,
   SelectingAction,
   MouseManager,
-  isNil,
+  DiagramModel,
+  EngineSetup,
+  KeyboardManager,
+  NodeModel,
+  CanvasManager,
+  ActionsManager,
 } from '@rxzu/core';
 import { ZonedClass, OutsideZone } from '../utils';
+import { EngineService } from '../engine.service';
+import { RegistryService } from '../registry.service';
+import { FactoryService } from '../factory.service';
+import { DIAGRAM_DEFAULT_OPTIONS } from '../injection.tokens';
 
 @Component({
-  selector: 'ngdx-diagram',
+  selector: 'rxzu-diagram',
+  exportAs: 'RxzuDiagram',
   templateUrl: 'diagram.component.html',
   styleUrls: ['diagram.component.scss'],
+  providers: [EngineService, RegistryService, FactoryService],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class RxZuDiagramComponent
-  implements AfterViewInit, OnDestroy, ZonedClass {
-  @Input('model') diagramModel: DiagramModel;
-  @Input() allowCanvasZoom = true;
-  @Input() allowCanvasTranslation = true;
-  @Input() inverseZoom = true;
-  @Input() allowLooseLinks = true;
-  @Input() maxZoomOut: number = null;
-  @Input() maxZoomIn: number = null;
-  @Input() portMagneticRadius = 30;
+export class RxZuDiagramComponent implements OnInit, OnDestroy, ZonedClass {
+  /** The name of the diagram, if not set will be `'default'` */
+  @Input() name?: string;
+  @Input() model!: DiagramModel;
+  @Input() options?: Partial<EngineSetup>;
 
-  @ViewChild('nodesLayer', { read: ViewContainerRef })
-  nodesLayer: ViewContainerRef;
+  @ViewChild('nodesLayer', { read: ViewContainerRef, static: true })
+  nodesLayer?: ViewContainerRef;
 
-  @ViewChild('linksLayer', { read: ViewContainerRef })
-  linksLayer: ViewContainerRef;
+  @ViewChild('linksLayer', { read: ViewContainerRef, static: true })
+  linksLayer?: ViewContainerRef;
 
-  @ViewChild('canvas', { read: ElementRef })
-  canvas: ElementRef;
+  @ViewChild('canvas', { read: ElementRef, static: true })
+  canvas?: ElementRef;
+  selectionBox$?: Observable<SelectingAction | null>;
 
-  diagramEngine: DiagramEngineCore;
   mouseManager: MouseManager;
-  protected destroyed$ = new ReplaySubject<boolean>(1);
-  protected selectionBox$: Observable<SelectingAction>;
+  keyboardManager: KeyboardManager;
+  canvasManager: CanvasManager;
+  actionsManager: ActionsManager;
 
   get host(): HTMLElement {
     return this.elRef.nativeElement;
   }
 
   constructor(
+    public readonly diagramEngine: EngineService,
     public ngZone: NgZone,
     protected renderer: Renderer2,
     protected cdRef: ChangeDetectorRef,
-    protected elRef: ElementRef<HTMLElement>
-  ) {}
+    protected elRef: ElementRef<HTMLElement>,
+    @Inject(DIAGRAM_DEFAULT_OPTIONS) protected defaultOptions: EngineSetup
+  ) {
+    this.mouseManager = this.diagramEngine.getMouseManager();
+    this.keyboardManager = this.diagramEngine.getKeyboardManager();
+    this.canvasManager = this.diagramEngine.getCanvasManager();
+    this.actionsManager = this.diagramEngine.getActionsManager();
+  }
 
-  ngAfterViewInit() {
-    if (this.diagramModel) {
-      this.diagramEngine = this.diagramModel.getDiagramEngine();
-      this.mouseManager = this.diagramEngine.getMouseManager();
-      this.diagramEngine.setCanvas(this.canvas.nativeElement);
-
-      this.diagramEngine.setup({
-        ...this,
+  ngOnInit() {
+    this.options = { ...this.defaultOptions, ...this.options };
+    this.model =
+      (this.model && this.diagramEngine.setModel(this.model)) ||
+      this.diagramEngine.createModel({
+        namespace: this.name || 'default',
+        ...this.options,
       });
 
-      (this.diagramEngine.paintNodes(this.nodesLayer) as Observable<boolean>)
-        .pipe(
-          switchMap(
-            () =>
-              this.diagramEngine.paintLinks(this.linksLayer) as Observable<void>
-          )
-        )
-        .subscribe(() => {
-          this.initSubs();
-          this.initSelectionBox();
-          this.cdRef.detectChanges();
-        });
+    if (!this.canvas) {
+      return;
     }
+
+    this.canvasManager.setCanvas(this.canvas.nativeElement);
+
+    this.diagramEngine.setup({
+      ...this.options,
+    } as EngineSetup);
+
+    this.initNodes();
+    this.initSelectionBox();
+    this.initSubs();
   }
 
   ngOnDestroy() {
-    this.destroyed$.next(true);
-    this.destroyed$.complete();
+    if (this.keyboardManager) {
+      this.keyboardManager.dispose();
+    }
+
+    this.model.destroy();
   }
 
-  initSelectionBox() {
-    this.selectionBox$ = this.diagramEngine.selectAction().pipe(
+  /**
+   * zoom the canvas to fit all nodes inside the view.
+   * @param additionalZoomFactor additional margins to the zooming factor
+   */
+  zoomToFit(additionalZoomFactor?: number): void {
+    this.diagramEngine.zoomToFit(additionalZoomFactor);
+  }
+
+  /**
+   *
+   * @param nodes zoom the canvas to the selected nodes
+   * @param margin additional margins to the zooming factor
+   */
+  zoomToNodes(nodes: NodeModel[], margin = 100): void {
+    this.diagramEngine.zoomToNodes(nodes, margin);
+  }
+
+  protected initNodes(): void {
+    this.canvasManager
+      .paintNodes(this.nodesLayer)
+      .pipe(
+        takeUntil(this.model.onEntityDestroy()),
+        switchMap(() => {
+          if (!this.diagramEngine) {
+            return of(null);
+          }
+
+          return this.canvasManager.paintLinks(this.linksLayer);
+        })
+      )
+      .subscribe();
+  }
+
+  protected initSelectionBox() {
+    if (!this.diagramEngine) {
+      return;
+    }
+
+    this.selectionBox$ = this.actionsManager.observeActions().pipe(
       map((a) => {
         if (
-          !isNil(a) &&
+          a &&
+          a.action &&
           a.action instanceof SelectingAction &&
           a.state === 'firing'
         ) {
-          return a.action;
+          return a.action as SelectingAction;
+        } else {
+          return null;
         }
-
-        return null;
       }),
       tap(() => this.cdRef.detectChanges())
-    ) as Observable<SelectingAction>;
+    );
   }
 
   @OutsideZone
-  onMouseUp(event: MouseEvent) {
-    this.mouseManager.onMouseUp(event);
+  protected onMouseUp(event: MouseEvent) {
+    this.mouseManager ? this.mouseManager.onMouseUp(event) : noop();
   }
 
   @OutsideZone
-  onMouseMove(event: MouseEvent) {
-    this.mouseManager.onMouseMove(event);
+  onKeyUp(event: KeyboardEvent) {
+    this.keyboardManager ? this.keyboardManager.onKeyUp(event) : noop();
+  }
+
+  @HostListener('window:copy', ['$event'])
+  protected onCopy(event: ClipboardEvent) {
+    this.keyboardManager ? this.keyboardManager.onCopy() : noop();
+  }
+
+  @HostListener('window:paste', ['$event'])
+  protected onPaste(event: ClipboardEvent) {
+    this.keyboardManager ? this.keyboardManager.onPaste() : noop();
+  }
+
+  @OutsideZone
+  protected onMouseMove(event: MouseEvent) {
+    this.mouseManager ? this.mouseManager.onMouseMove(event) : noop();
   }
 
   @OutsideZone
   onMouseDown(event: MouseEvent) {
-    this.mouseManager.onMouseDown(event);
+    this.mouseManager ? this.mouseManager.onMouseDown(event) : noop();
   }
 
   @OutsideZone
   onMouseWheel(event: WheelEvent) {
-    this.mouseManager.onMouseWheel(event);
+    this.mouseManager ? this.mouseManager.onMouseWheel(event) : noop();
   }
 
   @OutsideZone
@@ -144,23 +213,36 @@ export class RxZuDiagramComponent
   }
 
   protected initSubs() {
+    const diagramModel = this.diagramEngine?.getDiagramModel();
+    if (!diagramModel) {
+      return;
+    }
+
     combineLatest([
-      this.diagramModel.selectOffsetX(),
-      this.diagramModel.selectOffsetY(),
-      this.diagramModel.selectZoomLevel(),
+      diagramModel.selectOffsetX(),
+      diagramModel.selectOffsetY(),
+      diagramModel.selectZoomLevel(),
     ])
       .pipe(
-        tap(([x, y, zoom]) => this.setLayerStyles(x, y, zoom)),
-        takeUntil(this.destroyed$)
+        takeUntil(this.model.onEntityDestroy()),
+        tap(([x, y, zoom]) => this.setLayerStyles(x, y, zoom))
       )
       .subscribe();
   }
 
-  protected getNodesLayer(): HTMLDivElement {
-    return this.host.querySelector('.ngdx-nodes-layer');
+  protected getNodesLayer(): HTMLDivElement | null {
+    if (!this.host) {
+      return null;
+    }
+
+    return this.nodesLayer?.element.nativeElement.parentElement;
   }
 
-  protected getLinksLayer(): HTMLDivElement {
-    return this.host.querySelector('.ngdx-links-layer');
+  protected getLinksLayer(): HTMLDivElement | null {
+    if (!this.host) {
+      return null;
+    }
+
+    return this.linksLayer?.element.nativeElement.parentElement;
   }
 }
