@@ -2,49 +2,21 @@ import {
   ComponentFactoryResolver,
   Injectable,
   Renderer2,
-  RendererFactory2,
+  RendererFactory2
 } from '@angular/core';
-import { DiagramEngineCore } from '@rxzu/core';
-import {
-  DefaultLabelFactory,
-  DefaultLinkFactory,
-  DefaultNodeFactory,
-  DefaultPortFactory,
-} from './defaults';
-import { RxZuDiagramsModule } from './main.module';
+import { DiagramEngine } from '@rxzu/core';
+import { FactoryService } from './factory.service';
 
-@Injectable({ providedIn: RxZuDiagramsModule })
-export class DiagramEngine extends DiagramEngineCore {
+@Injectable()
+export class EngineService extends DiagramEngine {
   protected _renderer: Renderer2;
 
   constructor(
     protected resolver: ComponentFactoryResolver,
-    protected rendererFactory: RendererFactory2
+    protected rendererFactory: RendererFactory2,
+    protected factory: FactoryService
   ) {
-    super();
+    super(factory);
     this._renderer = this.rendererFactory.createRenderer(null, null);
-  }
-
-  registerDefaultFactories() {
-    const factoriesManager = this.getFactoriesManager();
-    factoriesManager.registerFactory({
-      type: 'nodeFactories',
-      factory: new DefaultNodeFactory(this.resolver, this._renderer),
-    });
-
-    factoriesManager.registerFactory({
-      type: 'linkFactories',
-      factory: new DefaultLinkFactory(this.resolver, this._renderer),
-    });
-
-    factoriesManager.registerFactory({
-      type: 'portFactories',
-      factory: new DefaultPortFactory(this.resolver, this._renderer),
-    });
-
-    factoriesManager.registerFactory({
-      type: 'labelFactories',
-      factory: new DefaultLabelFactory(this.resolver, this._renderer),
-    });
   }
 }
