@@ -1,11 +1,54 @@
 import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
+import {
+  Inject,
+  ModuleWithProviders,
+  NgModule,
+  Optional,
+  Provider,
+} from '@angular/core';
 import { RxZuDiagramComponent } from './diagram/diagram.component';
-import { TemplateVarDirective } from './utils/template-var.directive';
+import { TemplateVarDirective } from './utils';
+import { COMPONENT, ComponentProviderOptions } from './injection.tokens';
+import { RegistryService } from './registry.service';
+import { toRegistryKey } from '@rxzu/core';
 
 @NgModule({
   declarations: [RxZuDiagramComponent, TemplateVarDirective],
   imports: [CommonModule],
   exports: [RxZuDiagramComponent],
 })
-export class RxZuDiagramsModule {}
+export class RxZuModule {
+  constructor(
+    registry: RegistryService,
+    @Inject(COMPONENT) @Optional() components: ComponentProviderOptions[]
+  ) {
+    components?.forEach(({ type, namespace: name, component }) =>
+      registry.set(toRegistryKey(type, name), component)
+    );
+  }
+
+  static registerComponent({
+    type,
+    namespace,
+    component,
+  }: ComponentProviderOptions): Provider {
+    return {
+      provide: COMPONENT,
+      multi: true,
+      useValue: { type, namespace, component },
+    };
+  }
+
+  static withComponents(
+    components?: ComponentProviderOptions | ComponentProviderOptions[]
+  ): ModuleWithProviders<RxZuModule> {
+    return {
+      ngModule: RxZuModule,
+      providers: [
+        ...([] as ComponentProviderOptions[])
+          .concat(components || [])
+          .map(RxZuModule.registerComponent),
+      ],
+    };
+  }
+}
