@@ -4,11 +4,17 @@ export interface ZonedClass {
   ngZone: NgZone;
 }
 
-export function OutsideZone<T extends ZonedClass>(targetClass: T, functionName: string, descriptor) {
+export function OutsideZone<T extends ZonedClass>(
+  targetClass: T,
+  functionName: string,
+  descriptor: any
+) {
   const source = descriptor.value;
-  descriptor.value = function (...data) {
+  descriptor.value = function (...data: any[]) {
     if (!this.ngZone) {
-      throw new Error("Class with 'OutsideZone' decorator should have 'ngZone' class property with 'NgZone' class.");
+      throw new Error(
+        "Class with 'OutsideZone' decorator should have 'ngZone' class property with 'NgZone' class."
+      );
     }
     return this.ngZone.runOutsideAngular(() => source.call(this, ...data));
   };
