@@ -1,9 +1,8 @@
 export * from './lib/actions';
+export * from './lib/factories';
 export * from './lib/interfaces';
 export * from './lib/models';
 export * from './lib/utils';
-export * from './lib/factories';
-export * from './lib/defaults';
 export * from './lib/managers';
 export * from './lib/plugins';
 
