@@ -1,12 +1,7 @@
-export type BaseActionState = 'started' | 'firing' | 'stopped';
 export class BaseAction {
-  mouseX: number;
-  mouseY: number;
-  ms: number;
+    ms: number;
 
-  constructor(mouseX: number, mouseY: number) {
-    this.mouseX = mouseX;
-    this.mouseY = mouseY;
-    this.ms = new Date().getTime();
-  }
+    constructor() {
+        this.ms = new Date().getTime();
+    }
 }
