@@ -1,19 +1,19 @@
-import { BaseAction } from './base.action';
 import { PortModel } from '../models/port.model';
 import { NodeModel } from '../models/node.model';
 import { LinkModel } from '../models';
+import { BaseAction } from './base.action';
 
 export class InvalidLinkDestroyed extends BaseAction {
-  sourcePort: PortModel;
+  sourcePort: PortModel | null;
   link: LinkModel;
 
-  constructor(mouseX: number, mouseY: number, link: LinkModel) {
-    super(mouseX, mouseY);
-    this.sourcePort = link.getSourcePort();
+  constructor(link: LinkModel) {
+    super();
     this.link = link;
+    this.sourcePort = link.getSourcePort();
   }
 
-  getOutPortNode(): NodeModel {
-    return this.sourcePort.getParent();
+  getOutPortNode(): NodeModel | null {
+    return this.sourcePort?.getParent() ?? null;
   }
 }
