@@ -1,8 +1,7 @@
-import { BaseAction } from './base.action';
+import { BaseMouseAction } from './base-mouse.action';
 import { DiagramModel } from '../models/diagram.model';
-import { Coords } from '../interfaces/coords.interface';
 
-export class SelectingAction extends BaseAction {
+export class SelectingAction extends BaseMouseAction {
   mouseX2: number;
   mouseY2: number;
 
@@ -19,6 +18,14 @@ export class SelectingAction extends BaseAction {
     super(mouseX, mouseY);
     this.mouseX2 = mouseX;
     this.mouseY2 = mouseY;
+    this.dimensions = {
+      top: 0,
+      left: 0,
+      width: 0,
+      height: 0,
+      right: 0,
+      bottom: 0,
+    };
   }
 
   getBoxDimensions() {
@@ -28,20 +35,34 @@ export class SelectingAction extends BaseAction {
       width: Math.abs(this.mouseX2 - this.mouseX),
       height: Math.abs(this.mouseY2 - this.mouseY),
       right: this.mouseX2 < this.mouseX ? this.mouseX : this.mouseX2,
-      bottom: this.mouseY2 < this.mouseY ? this.mouseY : this.mouseY2
+      bottom: this.mouseY2 < this.mouseY ? this.mouseY : this.mouseY2,
     };
     return this.dimensions;
   }
 
-  containsElement({ x, y }: Coords, diagramModel: DiagramModel): boolean {
+  containsElement(
+    topLeftPoint: { x: number; y: number },
+    bottomRightPoint: { x: number; y: number },
+    diagramModel: DiagramModel
+  ): boolean {
     const z = diagramModel.getZoomLevel() / 100.0;
     const dimensions = this.getBoxDimensions();
 
-    return (
-      x * z + diagramModel.getOffsetX() > dimensions.left &&
-      x * z + diagramModel.getOffsetX() < dimensions.right &&
-      y * z + diagramModel.getOffsetY() > dimensions.top &&
-      y * z + diagramModel.getOffsetY() < dimensions.bottom
-    );
+    // check if box contain top left point
+    const isContainsTopLeftPoint =
+      topLeftPoint.x * z + diagramModel.getOffsetX() > dimensions.left &&
+      topLeftPoint.x * z + diagramModel.getOffsetX() < dimensions.right &&
+      topLeftPoint.y * z + diagramModel.getOffsetY() > dimensions.top &&
+      topLeftPoint.y * z + diagramModel.getOffsetY() < dimensions.bottom;
+
+    // check if box contain bottom right point
+    const isContainBottomRightPoint =
+      bottomRightPoint.x * z + diagramModel.getOffsetX() > dimensions.left &&
+      bottomRightPoint.x * z + diagramModel.getOffsetX() < dimensions.right &&
+      bottomRightPoint.y * z + diagramModel.getOffsetY() > dimensions.top &&
+      bottomRightPoint.y * z + diagramModel.getOffsetY() < dimensions.bottom;
+
+    // only if box contains both top left and bottom right points, the element is contains in the box
+    return isContainsTopLeftPoint && isContainBottomRightPoint;
   }
 }
