@@ -1,30 +1,44 @@
 import { MonoTypeOperatorFunction, Observable, Subject } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { BaseEvent, LockEvent } from './interfaces/event.interface';
-import { createValueState } from './state';
+import { createValueState, ValueState } from './state';
 import {
   entityProperty as _entityProperty,
   ID,
   log as _log,
   LOG_LEVEL,
   UID,
-  withLog as _withLog
+  withLog as _withLog,
 } from './utils/tool-kit.util';
 import { HashMap } from './utils/types';
-
-export type BaseEntityType = 'node' | 'link' | 'port' | 'point';
+import { BaseEntityOptions, BaseEntityType } from './interfaces';
 
 export class BaseEntity {
   protected _id: ID;
+  protected destroyed$: Subject<void>;
+  protected locked$: ValueState<boolean>;
+  protected namespace$: ValueState<string>;
 
-  protected destroyed$ = new Subject<void>();
-  protected locked$ = createValueState(false, this.entityPipe('locked'));
-
+  protected readonly _type: BaseEntityType;
   protected readonly _logPrefix: string;
 
-  constructor(id?: ID, logPrefix = '') {
-    this._id = id || UID();
-    this._logPrefix = `${logPrefix}`;
+  constructor(options: BaseEntityOptions) {
+    this._id = options.id || UID();
+    this._type = options.type;
+    this._logPrefix = `${options.logPrefix ?? ''}`;
+    this.destroyed$ = new Subject<void>();
+    this.locked$ = createValueState<boolean>(
+      !!options.locked,
+      this.entityPipe('locked')
+    );
+    this.namespace$ = createValueState<string>(
+      options.namespace ?? 'default',
+      this.entityPipe('name')
+    );
+  }
+
+  get type(): BaseEntityType {
+    return this._type;
   }
 
   get id(): ID {
@@ -33,6 +47,14 @@ export class BaseEntity {
 
   set id(id: ID) {
     this._id = id;
+  }
+
+  get namespace(): string {
+    return this.namespace$.value;
+  }
+
+  set namespace(value: string) {
+    this.namespace$.set(value);
   }
 
   log(message: string, ...args: any): void {
@@ -44,7 +66,11 @@ export class BaseEntity {
   }
 
   entityPipe<T>(logMessage = ''): MonoTypeOperatorFunction<T> {
-    return _entityProperty<T>(this.onEntityDestroy(), 0, `${this._logPrefix}: ${logMessage}`);
+    return _entityProperty<T>(
+      this.onEntityDestroy(),
+      0,
+      `${this._logPrefix}: ${logMessage}`
+    );
   }
 
   getLocked(): boolean {
@@ -52,11 +78,11 @@ export class BaseEntity {
   }
 
   setLocked(locked = true) {
-    this.locked$.set(locked).emit();
+    this.locked$.set(locked);
   }
 
   // eslint-disable-next-line
-  doClone(lookupTable: HashMap<any> = {}, clone: any) {
+  doClone(lookupTable: HashMap<any> = {}, clone: this) {
     /*noop*/
   }
 
@@ -72,13 +98,6 @@ export class BaseEntity {
 
     this.doClone(lookupTable, clone);
     return clone;
-  }
-
-  serialize() {
-    return {
-      id: this.id,
-      locked: this.getLocked()
-    };
   }
 
   lockChanges(): Observable<LockEvent> {
