@@ -1,6 +1,7 @@
 export * from './coords.interface';
 export * from './dimensions.interface';
 export * from './event.interface';
-export * from './serialization.interface';
+export * from './options.interface';
 export * from './move-selection.interface';
 export * from './select-options.interface';
+export * from './setup.interface';
