@@ -1,9 +1,3 @@
-export interface EngineSetup {
-  maxZoomIn?: number;
-  maxZoomOut?: number;
-  portMagneticRadius?: number;
-  allowLooseLinks?: boolean;
-  allowCanvasZoom?: boolean;
-  allowCanvasTranslation?: boolean;
-  inverseZoom?: boolean;
-}
+import { BaseEntityOptions, DiagramModelOptions } from './options.interface';
+
+export type EngineSetup = Omit<DiagramModelOptions, keyof BaseEntityOptions>;
