@@ -1,4 +1,4 @@
-import { BehaviorSubject, Observable, MonoTypeOperatorFunction } from 'rxjs';
+import { Observable, MonoTypeOperatorFunction } from 'rxjs';
 import { BaseEntity } from '../base.entity';
 import { EntityMap, ID } from '../utils';
 import { ValueState } from './value.state';
@@ -6,12 +6,8 @@ import { ValueState } from './value.state';
 export class EntityState<T extends BaseEntity> extends ValueState<
   EntityMap<T>
 > {
-  protected stream$: BehaviorSubject<EntityMap<T>>;
-
-  value$: Observable<EntityMap<T>>;
-
   constructor(
-    value?: EntityMap<T>,
+    value: EntityMap<T>,
     entityPipe?: MonoTypeOperatorFunction<EntityMap<T>>
   ) {
     super(value, entityPipe);
@@ -19,20 +15,16 @@ export class EntityState<T extends BaseEntity> extends ValueState<
 
   destroy() {
     this.clear();
-    this.stream$ = null;
-    this.value$ = null;
   }
 
-  clear(destroy = true) {
-    if (destroy) {
-      this.forEach((entity) => entity.destroy());
-    }
-
+  clear() {
+    this.forEach((entity) => entity.destroy());
     this.value.clear();
+    this.stream$.next(this.value);
     return this;
   }
 
-  get(id: ID): T {
+  get(id: ID): T | undefined {
     return this.value.get(id);
   }
 
@@ -42,6 +34,7 @@ export class EntityState<T extends BaseEntity> extends ValueState<
 
   add(entity: T): EntityState<T> {
     this.value.set(entity.id, entity);
+    this.stream$.next(this.value);
     return this;
   }
 
@@ -49,6 +42,7 @@ export class EntityState<T extends BaseEntity> extends ValueState<
     for (const entity of entities) {
       this.add(entity);
     }
+    this.stream$.next(this.value);
     return this;
   }
 
@@ -58,6 +52,7 @@ export class EntityState<T extends BaseEntity> extends ValueState<
     }
 
     this.value.delete(id);
+    this.stream$.next(this.value);
     return this;
   }
 
