@@ -1,8 +1,6 @@
 import { ID } from './tool-kit.util';
 
-export interface HashMap<T> {
-  [key: string]: T;
-}
+export type HashMap<T> = Record<string, T>;
 
 export type Entries<T, K = string> = (readonly [K, T])[];
 
